DEBUG Add logging to explore deadlock
Emit .fault log if db queue is suspended in
ArticlesTable.fetchArticles(fetchMethod:) as this seems to be the the nexus of the deadlock, based on backtraces.

Log application launch and background processing, to see if issue is related to suspend/background.

Add argument to .encodeWidgetData that's used to identify the caller in new log statement. The deadlock is occurring during widget data encoding and it's not obvious which of the call sites is the origin.